Move `BlockSquareMatrix` class to `operators`
